Improve internals of tensorprob
The 'Scalar' class has been renamed to 'Parameter' and is now just an
alias for 'Uniform'.

Internally, placeholders are now being rewritten as Variables when
model.observed is called.  The code for this should can probably be
cleaned up quite a bit and should be revisited at a later point.

I've also changed how 'model.pdf' works: It now assumes that it's inputs
(one for each observable) have the same shape. This might sound weird at
first, but is actually what we need for 'pdf', which is designed to help
visualize per-entry probabilities.

I had to comment out some tests, because I couldn't think of a way
to make them pass using the changed internals.
This should be revisited as well.